refactor(workflows): extract ForEachExpansionService and data_suffix
Move forEach iteration semantics out of WorkflowExecutionService into
a dedicated ForEachExpansionService. Pure transformation: takes a
job + an expression context, returns the map of expanded steps. No
I/O, no event emission, no orchestration concerns.

Three files now own what used to be inline in execution_service.ts:
- data_suffix.ts:        coerceToSuffix, MAX_SUFFIX_LENGTH
                         (used by both forEach iteration AND vary)
- for_each_expansion_service.ts: ForEachExpansionService class,
                         resolveForEachStepName, ExpandedStep,
                         ResolvedStepName
- execution_service.ts:  delegates via
                         `new ForEachExpansionService(new CelEvaluator())
                            .expand(job, expressionContext)`

Decoupling effect: bug fixes to forEach iteration semantics (array vs
object, fallback-suffix policy, step-name expression failure) no
longer risk affecting orchestration. The orchestrator stops knowing
how forEach works internally — it only consumes ExpandedStep[].

Both new files import zero infrastructure: ForEachExpansionService
takes a CelEvaluatorRuntime interface in its constructor (the
infrastructure CelEvaluator implements it structurally). DDD ratchet
unchanged at 26.

Behaviour-preserving:
- Same array-vs-object iteration, same fallback-suffix policy, same
  warning-log call sites with identical parameter substitution
- Same UserError messages
- Verified by /tmp/swamp-verification (10 phases, OK against baseline)
  and full deno test (4999 passed)
